requirements.txt
request
joblib
pandas
scikit-learn
streamlit
